requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-google-genai
tavily-python
langgraph
langchain-community